pyproject.toml
[project]
name = "icml-release"
version = "1.0.0"
description = "ICML 2026 - Mean Field Simulation Release"
authors = [
    {name = "ICML Team"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "torch (>=2.0.0,<3.0.0)",
    "transformers (>=4.30.0,<5.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "tensorboard (>=2.10.0,<3.0.0)",
    "matplotlib (>=3.5.0,<4.0.0)",
    "accelerate (>=0.20.0,<1.0.0)",
    "scipy (>=1.10.0,<2.0.0)",
    "huggingface-hub (>=0.16.0,<1.0.0)",
    "python-dotenv (>=0.19.0,<1.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "openai (>=1.0.0,<3.0.0)",
    "aiohttp (>=3.8.0,<4.0.0)",
    "scikit-learn (>=1.2.0,<2.0.0)",
    "wandb (>=0.24.0,<0.25.0)",
]

[tool.poetry]
name = "icml-release"
version = "1.0.0"
description = "ICML 2026 - Mean Field Simulation Release"
authors = ["ICML Team"]
readme = "README.md"
packages = [{include = "eval"}]

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
